switch to software volume scaling (it's bad tho i need daniel to do maths for me please)

+++ src/audio/include/bt_audio_output.hpp
@@ -54,7 +54,8 @@ class BluetoothAudioOutput : public IAudioOutput {
  private:
   drivers::Bluetooth& bluetooth_;
   tasks::WorkerPool& bg_worker_;
-  uint8_t volume_;
+
+  uint16_t volume_;
 };
 
 }  // namespace audio
